The outbuf for QP decoding can be up to len + 2 bytes and UU decoding
2001-12-05  Jeffrey Stedfast  <fejj@ximian.com>

	* gmime-filter-basic.c (filter_filter):
	(filter_complete): The outbuf for QP decoding can be up to len + 2
	bytes and UU decoding outbuf can be up to len + 3 bytes.

+++ gmime-stream-filter.c
@@ -185,7 +185,7 @@ stream_flush (GMimeStream *stream)
 	GMimeStreamFilter *filter = (GMimeStreamFilter *) stream;
 	struct _GMimeStreamFilterPrivate *p = filter->priv;
 	struct _filter *f;
-	int len, presize;
+	size_t len, presize;
 	char *buffer;
 	
 	if (p->last_was_read) {
